Initial commit - infrastructure for rust project
This includes a library `libpathfinder` and a binary `charsheet`. The
library is intended to be a generic useful pathfinder library with all
of the pathfinder constructs. The binary is intended to serve as a
character sheet for a played character in a base-rules pathfinder
campaign.

This commit only includes a character with a name, that can be
serialized to json. The rest to come™.