Use Html.Program as main for elm app
This allows us to retrieve data and update the app with its contents.
This will certainly evolve over time, but starting with a skeleton of
something useful here.

(*) See http://elm-lang.org/examples/http for the reference this was
based off of.

(*) http://package.elm-lang.org/packages/elm-lang/core/5.1.1/Json-Decode
is a useful resource as well.