Introduce sqlite interactions
This includes the bootstrapping methodology, as well as the code change
to use this sqlite data for a very simple character-retrieval. Going to
polish this interaction in the datastore layer where I'm hoping to have
cleaner definitions of interactions with the canonical models without
manually mapping row results to constructors.

The bootstrapping is as simple as running the `seed.sh` script in the
`src/libpathfinder/db/` directory. This will create the `local_pf.db`
file that the app is currently hard coded to read from.

Also including instructions in the README for the `src/pathserver`
directory on how to bootstrap the database.