Also move webshared to libpathfinder_common
This is needed by the macro derive, but was working with the only
call-point having access to this module already. This should make the
dependency on the macro only be the extern crates `iron`, `urlencoded`,
and `libpathfinder_common`, and the macro should not depend on
existing `use`s in the module it is expanded in.