Derive macro for FromDB for arbitrary structs
This will allow use to simply:

    #[derive(FromDB)]
    #[from_db(table_name = "characters")]
    pub struct Character { ... }

for all structs we add to map to tables. The field names will have to
match the table field names, and the table_name attribute must be
accurate, but with those true, we will be able to save a lot of
duplication with the sql construction and construction of the struct
with those fields.