Separate canonical character from db representation
As the character struct gets more complex, the database representation
will not directly lend itself to what we will want the canonical model
to be. Keeping the canonical struct the same as the web representation,
as that public view of the character struct will be synonymous with the
canonical model.