Introduce Postgresql interactions
After not being able to alter columns in sqlite, decided to bite the
bullet and prefer postgresql for the database. These are the scripts and
new schema/data for bootstrapping a new database. The README.md includes
hints for how to get postgresql up and running on a developer machine.