Convert rust code from rusqlite -> rust-postgres
This changes in a few key places, but is relatively isolated with the
changes, the custom FromSql trait for the alignment enums has to change
to match rust-postgres's pattern, and the derive macro for our custom
FromDB trait also has to be changed. Other than that most of the changes
are pretty mechanical rusqlite -> rust-postgres replacements.