Have proper error/loading handling in elm app
This is the first step in allowing different data than just a character
to be loaded, and was overdue with the error handling that we already
had to build. Would like to build the error json handling next, going to
remove some now dead code, wanted this transformation to be as clean as
possible.