Add player name
This, as an exercise of how much effort it is to add a new field, is
pretty promising. Not displaying this in the prettiest of ways, will
rework that in the near-future. Will probably also put this
lore-but-not-number-crunch-y data in a separate sub-struct of Character,
for the canonical model (on the rust end, as well as the elm end). It
seems elm's Json.Decode.map9 being the max may prove to be a problem
otherwise.