Include class sub skills in skill information
These are only different from the basic skills in that they reference
the sub_skills table, but otherwise operate effectively the same.

We do need to include the sub_skills_id in the augmented character's sub
skill choices, so that we can properly look it up in turning the
underlying data into the canonical model for a character.