Include class skill constructors for class skill perks
These perks apply to all sub skills you pick in a given skill
constructor:

https://rpg.stackexchange.com/questions/19869/does-every-specialization-of-craft-you-take-get-the-class-skill-bonus-or-just-t

The key takeaways being:

Quotes
======

Q:
Suppose I'm playing a character with Craft as a class skill. I put 1
rank in it, select Weaponsmith to fit my background and end up with
Craft (Weapons) +4. Now, if I later decide I want to be able to make
bows and arrows, do I also get another +3 bonus on Craft (Bows) when I
invest a new rank? There are different kinds of craft, just like there
are different kinds of knowledge, but the class skill bonus for craft is
completely unspecified.

A:
Yes.
Craft is actually a number of separate skills. So if you have the class
bonus for craft you have it for all the skills in the craft "family".

A:
If you only had 1 rank in weapons and not another in bows, it only
affects that specific craft skill. If you are spending 2 ranks in them,
1 in both, then they are both receiving the bonus for being a class
skill.
(I think this answer emphasises something missing in the first: that you
don't get the +3 bonus in another craft until you spend a point in that
specialty.)

Extra Notes
===========

The logic that I don't have covered here is that you need to select the
craft a given number of times. I wouldn't normally say I have Craft(x)
until I put the point in x, so I will leave this as an open bug for now.
In fact, I don't even allow you to mark sub_skills without selecting a
rank in them, so this bug will be attached to that.