Sane constructor for tables
This takes tuples of headers and values, which works for the meta
information, as there is only one row. I may need to take Html elements
instead, to allow for varying display amongst columns. I will also need
to allow it to take lists of display rows, for all of the tables that
have more than one row.