Include weapon(s) on creatures (and character sheet)
A similar mapping mechanism is in place here, but we are allowing for
multiple weapons in this case, as multiple weapons can be owned and used
by a single creature/character. May add a "equipped weapon" field, but
may leave that to the player to keep track of in their head for the time
being.