Introduce class armor proficiency
This is a table that holds the reduction in armor check penalty, as well
as the increase in max dexterity. In my case, I'm not pushing the max
dex limit, so still deferring on implementing that. This change
includes:
    * introducing the new table for this data, which keeps track of the
      check penalty reduction, and max dex for a given class/level combo
    * introducing the rust struct to match said table
    * querying this table when answering the character api endpoint, and
    * using this data when building the armor check penalty value used
      across the skill table

We do some small refactoring to calculate this armor check penalty for
skills once, and then reference this value when applying it to skills.