Add spiked shield information
This builds enough of the requisite pieces to query for this data and
present it to the frontend. We're storing the spiked shield information
on the mapping table from characters to shields, instead of introducing
another entry to point to. We are preferring not to put the spiked
information on the shields themselves, so that we don't have to
duplicate each shield, for the spiked and non-spiked version.