Prefer to operate on models.Weapon in combat conversion
Instead of operating on the db struct, preferring the canonical model.
This will allow shields to convert to canonical model Weapons and then
be able to convert to the combat representation. Extracting both of
these up, as we do need to compute the full_weapon to pass to the
build_combat_weapon_stats, so extracting both up for clarity.