Endpoint to POST creature item
This includes primitive insertion logic and method switching. We have
this handler only insert on POST with a switch on the HTTP method, and
return a method not allowed if it receives any other method. The
insertion logic is not yet using traits or derives, that will come as we
have more insertion.